requirements.txt
aiokafka==0.7.2
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
elastic-transport==8.4.0
elasticsearch==8.8.0
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
influxdb-client==1.32.0
iniconfig==1.1.1
kafka-python==2.0.2
loguru==0.6.0
packaging==21.3
pluggy==1.0.0
prometheus-client==0.14.1
py==1.11.0
pycodestyle==2.9.1
pydantic==1.10.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
reactivex==4.0.4
redis==4.3.4
requests==2.31.0
requests-toolbelt==0.9.1
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websocket-client==1.4.0
websockets==10.3
wrapt==1.14.1
